+++ sdk/resourcemanager/azure-resourcemanager-appservice/src/test/java/com/azure/resourcemanager/appservice/FunctionAppsTests.java
@@ -10,6 +10,7 @@ import com.azure.core.management.Region;
 import com.azure.core.management.exception.ManagementException;
 import com.azure.core.management.profile.AzureProfile;
 import com.azure.resourcemanager.appcontainers.ContainerAppsApiManager;
+import com.azure.resourcemanager.appcontainers.models.ManagedEnvironment;
 import com.azure.resourcemanager.appservice.models.AppServicePlan;
 import com.azure.resourcemanager.appservice.models.AppSetting;
 import com.azure.resourcemanager.appservice.models.FunctionApp;
@@ -19,18 +20,19 @@ import com.azure.resourcemanager.appservice.models.FunctionEnvelope;
 import com.azure.resourcemanager.appservice.models.FunctionRuntimeStack;
 import com.azure.resourcemanager.appservice.models.PricingTier;
 import com.azure.resourcemanager.appservice.models.SkuName;
-import com.azure.resourcemanager.resources.fluentcore.arm.ResourceId;
 import com.azure.resourcemanager.resources.fluentcore.utils.ResourceManagerUtils;
 import com.azure.resourcemanager.resources.models.ResourceGroup;
 import com.azure.resourcemanager.storage.StorageManager;
 import com.azure.resourcemanager.storage.models.StorageAccount;
 import com.azure.resourcemanager.storage.models.StorageAccountSkuType;
+import com.azure.resourcemanager.test.utils.TestDelayProvider;
 import com.azure.resourcemanager.test.utils.TestUtilities;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
+import java.lang.reflect.Constructor;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Map;
@@ -61,7 +63,21 @@ public class FunctionAppsTests extends AppServiceTest {
         rgName2 = generateRandomResourceName("javacsmrg", 20);
 
         storageManager = buildManager(StorageManager.class, httpPipeline, profile);
-        containerAppsApiManager = ContainerAppsApiManager.authenticate(httpPipeline, profile);
+        // build ContainerAppsApiManager
+        try {
+            Constructor<ContainerAppsApiManager> constructor = ContainerAppsApiManager.class.getDeclaredConstructor(httpPipeline.getClass(), profile.getClass(), Duration.class);
+            Runnable runnable = () -> constructor.setAccessible(true);
+            runnable.run();
+            ResourceManagerUtils.InternalRuntimeContext.setDelayProvider(new TestDelayProvider(!isPlaybackMode()));
+            containerAppsApiManager = constructor.newInstance(
+                httpPipeline,
+                profile,
+                // Lite packages need this for playback mode, since they take defaultPollInterval directly as polling interval.
+                ResourceManagerUtils.InternalRuntimeContext.getDelayDuration(Duration.ofSeconds(30)));
+
+        } catch (ReflectiveOperationException ex) {
+            throw LOGGER.logExceptionAsError(new RuntimeException(ex));
+        }
 
         super.initializeClients(httpPipeline, profile);
     }
@@ -479,20 +495,20 @@ public class FunctionAppsTests extends AppServiceTest {
             .define(rgName1)
             .withRegion(region)
             .create();
-        String managedEnvironmentName = createAcaEnvironment(region, resourceGroup);
+        String managedEnvironmentId = createAcaEnvironment(region, resourceGroup);
         webappName1 = generateRandomResourceName("java-function-", 20);
         FunctionApp functionApp = appServiceManager
             .functionApps()
             .define(webappName1)
             .withRegion(region)
             .withExistingResourceGroup(resourceGroup)
-            .withManagedEnvironmentName(managedEnvironmentName)
+            .withManagedEnvironmentId(managedEnvironmentId)
             .withMaxReplicas(10)
             .withMinReplicas(3)
             .withPublicDockerHubImage("mcr.microsoft.com/azure-functions/dotnet7-quickstart-demo:1.0")
             .create();
 
-        Assertions.assertEquals(managedEnvironmentName, ResourceId.fromString(functionApp.managedEnvironmentId()).name());
+        Assertions.assertEquals(managedEnvironmentId, functionApp.managedEnvironmentId());
         Assertions.assertEquals(10, functionApp.maxReplicas());
         Assertions.assertEquals(3, functionApp.minReplicas());
 
@@ -507,13 +523,13 @@ public class FunctionAppsTests extends AppServiceTest {
 
     private String createAcaEnvironment(Region region, ResourceGroup resourceGroup) {
         String managedEnvironmentName = generateRandomResourceName("jvacam", 15);
-        containerAppsApiManager.managedEnvironments()
+        ManagedEnvironment managedEnvironment = containerAppsApiManager.managedEnvironments()
             .define(managedEnvironmentName)
             .withRegion(region)
             .withExistingResourceGroup(resourceGroup.name())
             .withZoneRedundant(false)
             .create();
-        return managedEnvironmentName;
+        return managedEnvironment.id();
     }
 
     private void assertRunning(FunctionApp functionApp) {